% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAS_data_modify_types.R
\name{as_numeric_RC}
\alias{as_character_OT}
\alias{as_numeric_RC}
\alias{to_factor_RS}
\title{ISAS data manipulation tools}
\usage{
as_numeric_RC(df)

as_character_OT(df)

to_factor_RS(df)
}
\arguments{
\item{df}{ISAS data.}
}
\description{
ISAS data manipulation tools
}

