% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EOL_DP_tools.r
\name{combine_pipe_table}
\alias{combine_pipe_table}
\title{Combine same choices from ISAS "table + piping + multiple choice"}
\usage{
combine_pipe_table(.data, .var_start, .var_end, .MR_num, .step)
}
\arguments{
\item{.data}{Data.}

\item{.var_start}{Dariable name of table start.}

\item{.var_end}{Variable name of table end.}

\item{.MR_num}{Number of items in a multiple response question.}

\item{.step}{Number of duplicated questions in a question set.}
}
\description{
Combine same choices from ISAS "table + piping + multiple choice"
}

