% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{compare_sample_size}
\alias{compare_sample_size}
\title{Compare sample size}
\usage{
compare_sample_size(.data, .strataCrossN_flat, .vars, .namesList)
}
\arguments{
\item{.data}{}

\item{.strataCrossN_flat}{}

\item{.vars}{}

\item{.namesList}{}
}
\description{
Compare sample size
}

