% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copy_paste_clipboard.R
\name{copy_tbl}
\alias{copy_tbl}
\alias{paste_tbl}
\title{Copy Data out of or Paste data into R}
\usage{
copy_tbl(obj, size = 4096)

paste_tbl()
}
\arguments{
\item{obj}{object to copy}

\item{size}{size of memory}
}
\description{
Using the clipboard, or passing data quickly from Excel to R
and back again.
}
\examples{
# These all work
copy_tbl(1:100)
copy_tbl(letters)
copy_tbl(iris)
copy_tbl(table(iris$Species))
copy_tbl(matrix(1:20, nrow = 2))

\dontrun{
# If my.df is of moderate size
copy_tbl(my.df)

# If my.df is huge
copy_tbl(my.df, 10000)

# Pasting works in a similar way. Select the range in Excel you want to copy
# (including the header) and press Ctrl+C. Then run

other.df <- paste_tbl()
}
}
\references{
Using the Windows Clipboard, or Passing Data Quickly From Excel to R and Back Again

http://www.r-bloggers.com/using-the-windows-clipboard-or-passing-data-quickly-from-excel-to-r-and-back-again/

http://stackoverflow.com/questions/9035674/r-function-to-copy-to-clipboard-on-mac-osx
}

