% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eol_report_crawler.R
\name{eol_report_crawler}
\alias{eol_report_crawler}
\title{Get Panel Supply IDs report}
\usage{
eol_report_crawler(survey_id)
}
\arguments{
\item{survey_id}{ISAS survey id.}
}
\value{
a \code{tbl_df} object
}
\description{
Get Panel Supply IDs report
}
\examples{
\dontrun{
eol_report_crawler(3896)
}
}

