% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EOL_DP_tools.r
\name{fix_pipe_table}
\alias{fix_pipe_table}
\title{Fix ISAS table + piping + multiple choice}
\usage{
fix_pipe_table(.data, .var_start, .var_end, .MR_num)
}
\arguments{
\item{.data}{Data frame to fix.}

\item{.var_start}{Variable name of table start.}

\item{.var_end}{Variable name of table end.}

\item{.MR_num}{Number of items in a multiple response question.}
}
\description{
Fix ISAS table + piping + multiple choice
}

