% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_id.r
\name{gen_fake_id}
\alias{gen_fake_id}
\title{Export fake IDs}
\usage{
gen_fake_id(ids, path, key = c("p", "j"), .survey_id = NULL,
  .outurl = NULL, .etc1 = NA)
}
\arguments{
\item{ids}{Real IDs.}

\item{path}{.xlsx fake id file path to export.}

\item{key}{Either "p" or "j" for fake IDs.}

\item{.survey_id}{ISAS survey id.}

\item{.outurl}{Survey URL provided by client.}

\item{.etc1}{etc1 column value.}
}
\value{
NULL. Generate file at provided path.
}
\description{
Export fake IDs
}
\examples{
\dontrun{
id_output \%>>\%
## Export pid
(~ gen_fake_id(., path, .survey_id = survey_id, .outurl = outurl)) \%>\%
## Export mail ids
  write_mail_list(mail_path)
}
}

