% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{margin_prop}
\alias{margin_prop}
\title{Check marginal proportion}
\usage{
margin_prop(.strataCrossN_flat, export = FALSE,
  outputfile = "R<U+00B9><ea><U+00A9><e2><U+00BC><U+02E5><U+00BB>temp.csv")
}
\arguments{
\item{.strataCrossN_flat}{}

\item{export}{}

\item{outputfile}{}
}
\description{
Check marginal proportion
}

