% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_id_crawler.R
\name{panel_id_crawler}
\alias{panel_id_crawler}
\title{panel id crawler}
\usage{
panel_id_crawler()
}
\description{
save .csv file to system.file("extdata/panel_data.csv", package = "EOLembrainToolbox")
}

