% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ids.r
\name{read_ids}
\alias{read_ids}
\title{Read IDs from single txt file path, or txt files in directory}
\usage{
read_ids(path)
}
\arguments{
\item{path}{txt file path, or directory contains txt files}
}
\value{
character vector.
}
\description{
Read IDs from single txt file path, or txt files in directory
}

