% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_panel_id.r
\name{read_panel_id}
\alias{read_panel_id}
\title{Read panel IDs}
\usage{
read_panel_id(path = c("lib_path", "update"))
}
\arguments{
\item{path}{Either "lib_path" for built-in data, or "update" to update first.}
}
\value{
\code{tbl_df} object of panel data
}
\description{
Read panel IDs
}

