% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_to.r
\name{recode_to}
\alias{recode_to}
\title{Recode variables}
\usage{
recode_to(df, ..., match = c("first", "last"))
}
\arguments{
\item{df}{A data.frame object}

\item{...}{Comma separated list of unquoted expressions. See examples.}

\item{match}{Variables matching  sequence. "first" or "last"}
}
\description{
Recode variables
}
\examples{
recode_to(mtcars,
  mpg < 20 ~ "a",
  vs == 1  ~ "b"
)

## empty lhs for "else" condition
recode_to(mtcars,
  mpg < 20 ~ "a",
  vs == 1  ~ "b",
           ~ "c"
)
recode_to(mtcars,
  mpg < 20 ~ mpg,
           ~ mpg + 100
)
}
\references{
forked from hadley/recode.R
https://gist.github.com/hadley/2751ba61d1c7f4eaacab
}

