% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ids.r, R/sample_ids.r
\name{sample_N}
\alias{sample_N}
\title{Sampling from panel data}
\usage{
sample_N(.data, .n, .id_var = "Panel_id", sent_id = NULL,
  finished_id = NULL, include_sent = FALSE, show = TRUE)

sample_N(.data, .n, .id_var = "Panel_id", sent_id = NULL,
  finished_id = NULL, include_sent = FALSE, show = TRUE)
}
\arguments{
\item{.data}{Panel data generated by \code{\link{read_panel_id}()}.}

\item{.n}{Number of samples to be extracted}

\item{.id_var}{Default "Panel_id".}

\item{sent_id}{}

\item{finished_id}{}

\item{include_sent}{}

\item{show}{}

\item{df}{Panel data generated by \code{\link{read_panel_id}()}.}

\item{.n}{Number of samples to be extracted}

\item{.id_var}{Default "Panel_id".}

\item{sent_id}{}

\item{finished_id}{}

\item{include_sent}{}

\item{show}{}
}
\value{
character vector of extracted IDs.

character vector of extracted IDs.
}
\description{
Sampling from panel data

Sampling from panel data
}
\examples{
sent_id <- EOLembrainToolbox::read_ids("./exclude_id")
finished_id <- c(EOLembrainToolbox::read_ids("./exclude_id/finished_ID"),
                 EOLembrainToolbox::eol_report_crawler(survey_id)$Panel_ID)
panel_id_all <- read_panel_id()
panel_id <- panel_id_all \%>\% filter(active_grade \%in\% 1:2)  # active, semi-active
panel_id \%>\%
filter(gender \%in\% 1, age \%in\% 20:29) \%>\%
  sample_N(2000, sent_id = sent_id, finished_id = finished_id, include_sent = T)


sent_id <- EOLembrainToolbox::read_ids("./exclude_id")
finished_id <- c(EOLembrainToolbox::read_ids("./exclude_id/finished_ID"),
                 EOLembrainToolbox::eol_report_crawler(survey_id)$Panel_ID)
panel_id_all <- read_panel_id()
panel_id <- panel_id_all \%>\% filter(active_grade \%in\% 1:2)  # active, semi-active
panel_id \%>\%
filter(gender \%in\% 1, age \%in\% 20:29) \%>\%
  sample_N(2000, sent_id = sent_id, finished_id = finished_id, include_sent = T)


}

