% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{sampleC}
\alias{sampleC}
\title{sample C}
\usage{
sampleC(.data, .strataCrossN_flat, .vars, .namesList)
}
\arguments{
\item{.data}{}

\item{.strataCrossN_flat}{}

\item{.vars}{}

\item{.namesList}{}
}
\description{
sample C
}

