% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_outeater.r
\name{sampling_outeater}
\alias{sampling_outeater}
\title{Outeater Sampling}
\usage{
sampling_outeater(panel_id = panel_id_active, sampleN, sent_id = NULL,
  finished_id = NULL)
}
\arguments{
\item{panel_id}{}

\item{sampleN}{}

\item{sent_id}{}

\item{finished_id}{}
}
\value{
character vector
}
\description{
Outeater Sampling
}
\examples{
sampleN <- list(
TP = list(N = 20, age = c(35:59), gender=1:2, include_sent=F),
NewTP = list(N = 50, age = c(15:19, 35:59), gender=1:2, include_sent=F),
KL_YL = list(N = 30, age = c(15:59), gender=1:2, include_sent=T),
TY_XC_ML = list(N = 100, age = c(15:20, 35:59), gender=1:2, include_sent=F),
TCH_CHW_NT = list(N = 130, age = c(15:20, 35:59), gender=1:2, include_sent=F),
YL_CHY_TN = list(N = 100, age = c(15:25, 35:59), gender=1:2, include_sent=F),
KS_PT = list(N = 100, age = c(15:20, 35:59), gender=1:2, include_sent=F),
HWL_TD = list(N = 20, age = c(15:49), gender=1:2, include_sent=T)
)

id_output <- panel_id_all \%>\%
filter(active_grade \%in\% 1:2) \%>\%  # active, semi-active
  sampling_outeater(sampleN,
                    sent_id,
                    finished_id)


}

