% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/syntax_crawler.R
\name{syntax_crawler}
\alias{syntax_crawler}
\title{SPSS syntax crawler}
\usage{
syntax_crawler(survey_id)
}
\arguments{
\item{survey_id}{}
}
\description{
SPSS syntax crawler
}
\examples{
syntax_crawler(3913)
}

