% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{to_strataCrossN}
\alias{to_strataCrossN}
\title{Create strataCrossN}
\usage{
to_strataCrossN(.strataPctList, .namesList, N)
}
\arguments{
\item{.strataPctList}{See example.}

\item{.namesList}{See example.}

\item{N}{Total sample size.}
}
\description{
Create strataCrossN
}
\examples{
N = 900

strataPctList <- list(
  sex = c(0.5, 0.5),
  age = c(1/3, 1/3, 1/3),
  area = c(0.5, 0.2, 0.3)
)

namesList <- list(sex=1:2, age=2:4, area=1:3)
vars <- c("Q1R", "Q2R", "Q3R")   # names(namesList)

strataCrossN_flat <- as.strataCrossN(strataPctList, namesList, N)
}

