% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_character_string.R
\name{V}
\alias{V}
\alias{V_tab}
\title{Paste character string from clipboard into R}
\usage{
V(addC = TRUE)

V_tab(addC = TRUE)
}
\arguments{
\item{addC}{add \code{c(...)} or not}
}
\description{
Paste character string from clipboard into R
}
\details{
\code{V}() for column of string vector

\code{V_tab}() for tab seperated string vector
}
\examples{
## Typical application:

# Step 1: paste a column of variable names from a meta database;  
# Step 2: Run this function in R: V()  
# Step 3: Press control(command) + V: Paste vector of clipboard  


}
\seealso{
see
\url{http://www.r-bloggers.com/efficient-variable-selection-in-r/}
}

