% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_mail_list.r
\name{write_mail_list}
\alias{write_mail_list}
\title{write_mail_list}
\usage{
write_mail_list(ids, file_name)
}
\arguments{
\item{ids}{}

\item{file_name}{}
}
\value{
txt file
}
\description{
write_mail_list
}
\examples{
\dontrun{
id_output \%>\% write_mail_list("xxxx_sent_id.txt")
}
}

