% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{write_output_csv}
\alias{write_output_csv}
\title{write_output_csv}
\usage{
write_output_csv(ID_sample, df, ID_var_name)
}
\arguments{
\item{ID_sample}{Character vector of IDs.}

\item{ID_var_name}{}
}
\description{
write_output_csv
}

