% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOL_DP_tools.r
\name{combine_pipe_table}
\alias{combine_pipe_table}
\title{Combine same choices from ISAS "table + piping + multiple choice"}
\usage{
combine_pipe_table(df, var_start, var_end, mr_C_max, steps, rep_place = c("q",
  "c"))
}
\arguments{
\item{df}{Data.}

\item{var_start}{Dariable name of table start.}

\item{var_end}{Variable name of table end.}

\item{mr_C_max}{Number of items in a multiple response question, excludes repeated.}

\item{steps}{Number of duplicated questions in a questionset.}

\item{rep_place}{Where the repeated items locate.
"q" for question; "c" for choices}
}
\value{
data.frame result
}
\description{
Combine same choices from ISAS "table + piping + multiple choice"
}

