% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISAS_data_DP_tools.r
\name{fix_pipe_table}
\alias{fix_pipe_table}
\title{Fix ISAS table + piping + multiple choice}
\usage{
fix_pipe_table(df, var_start, var_end, mr_C_max)
}
\arguments{
\item{df}{Data frame to fix.}

\item{var_start}{Variable name of table start.}

\item{var_end}{Variable name of table end.}

\item{mr_C_max}{Number of items in a multiple response question.}
}
\description{
Fix ISAS table + piping + multiple choice
}

