% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{margin_prop}
\alias{margin_prop}
\title{Check marginal proportion}
\usage{
margin_prop(.strataCrossN_flat, export = FALSE,
  outputfile = "R實抽樣本temp.csv")
}
\arguments{
\item{.strataCrossN_flat}{}

\item{outputfile}{}
}
\description{
Check marginal proportion
}

