% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_panel_id.r
\name{read_panel_id}
\alias{read_panel_id}
\title{Read panel IDs}
\usage{
read_panel_id(file = "lib_path", update = FALSE)
}
\arguments{
\item{update}{\code{TRUE} to update built-in data first}

\item{path}{Either "lib_path" for built-in data, or csv file path.}
}
\value{
\code{tbl_df} object of panel data
}
\description{
Read panel IDs
}

