% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{strata_ids}
\alias{strata_ids}
\title{strata sample ID}
\usage{
strata_ids(.data, .strataCrossN_flat, .namesList, .vars,
  ID_var_name = "Panel.ID", method = c("srswor", "srswr", "poisson",
  "systematic"), pik = NULL)
}
\arguments{
\item{.data}{}

\item{.strataCrossN_flat}{}

\item{.namesList}{}

\item{.vars}{}

\item{ID_var_name}{}

\item{method}{}

\item{pik}{}
}
\description{
strata sample ID
}

