% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_sampling.r
\name{write_output_xlsx}
\alias{write_output_xlsx}
\title{write output xlsx}
\usage{
write_output_xlsx(ID_sample, .data, ID_var_name, fileName)
}
\arguments{
\item{fileName}{}
}
\description{
write output xlsx
}

